+++ src/main/java/seedu/address/model/TodoList.java
@@ -5,8 +5,8 @@ import static java.util.Objects.requireNonNull;
 import java.util.List;
 
 import javafx.collections.ObservableList;
-import seedu.address.model.todolist.Task;
-import seedu.address.model.todolist.UniqueTodoList;
+import seedu.address.model.task.Task;
+import seedu.address.model.task.UniqueTodoList;
 
 /**
  * Wraps all data at the todo list level.
