+++ src/main/java/seedu/address/model/task/exceptions/DuplicateTaskException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.todolist.exceptions;
+package seedu.address.model.task.exceptions;
 
 /**
  * Signals that the operation will result in duplicate Tasks (Tasks are considered duplicates if they have the same
