+++ src/main/java/seedu/address/model/task/exceptions/TaskNotFoundException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.todolist.exceptions;
+package seedu.address.model.task.exceptions;
 
 /**
  * Signals that the operation is unable to find the specified task;
